% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{as.matrix.bmixlm}
\alias{as.matrix.bmixlm}
\title{Convert \code{bmixlm} object to a matrix}
\usage{
\method{as.matrix}{bmixlm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bmixlm}.}

\item{...}{Currently ignored.}
}
\value{
A matrix of samples of model parameters.
}
\description{
Returns all the samples of the parameters as one large matrix.
}
