% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm-package.R
\docType{package}
\name{bmixlm-package}
\alias{bmixlm-package}
\title{bmixlm: Binary Mixtures of Linear Models}
\description{
Fit a binary mixture of linear models by Gibbs sampling. Assumes responses are drawn two subpopulations, each described by a different linear model, and it is unknown which responses are drawn from which subpopulation. The mixture model estimates the coefficient of the two subpopulation models and the probability that an observation is drawn from a particular subpopulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/bmixlm}
  \item Report bugs at \url{https://github.com/SWotherspoon/bmixlm/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
