% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{bmixlm}
\alias{bmixlm}
\title{Binary Mixture of Linear Models}
\usage{
bmixlm(formula1, formula2, formulap, data, nsamp = 1000, nthin = 3,
  tau.prior = c(0.01, 0.01), start = list(sigma = c(1e-04, 1e-04)))
}
\arguments{
\item{formula1}{An object of class formula: a symbolic description
of the linear model for the first component}

\item{formula2}{An object of class formula: a symbolic description
of the linear model for the second component}

\item{formulap}{An object of class formula: a symbolic description
of the probit model for the probability an observation is}

\item{data}{A dataframe containing the variables from the model.}

\item{nsamp}{Number of samples to draw.}

\item{nthin}{The thinning rate.}

\item{tau.prior}{Parameters of the (common) Gamma prior for the
precision of the two models.}

\item{start}{A list of initial values for sigma, betap and b}
}
\value{
An object of class \code{bmixlm} with columns
\item{call}{the matched call}
\item{nsamp}{the number of samples retained after thinning}
\item{beta1}{matrix of samples of the coefficients of the first
linear model}
\item{beta2}{matrix of samples of the coefficients of the second
linear model}
\item{betap}{matrix of samples of the coefficients of the probit
model}
\item{sigma}{two column matrix of samples of the standard
deviations of the errors for the two models}
\item{data}{the input dataframe}
\item{pW}{effective degrees of freedom for the fitted model}
\item{WAIC}{WAIC for the fitted model}
\item{restart}{final sigma, betap and b for restart purposes}
}
\description{
Fits a binary mixture of linear models in which the probability of
class membership is related to the covariates through a probit
regression model.
}
\details{
The model assumes the observations are drawn from a two component
mixture, where each component is described by a different linear
model.  The probability that an individual observation is a member
of one component or the other is modelled by a probit regression.

The model is fit by Gibbs sampling, assuming uniform priors for
the regression coefficients of the two linear model and the probit
regression, and a (common) Gamma prior for the precision (inverse
variance) of the two linear models.  The probit component of the
model is sampled by the method of Albert and Chib.
}
\references{
Albert, J. H., & Chib, S. (1993). Bayesian analysis of
  binary and polychotomous response data. Journal of the American
  statistical Association, 88(422), 669-679.
}

