% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{coef.bmixlm}
\alias{coef.bmixlm}
\title{Coefficients of a \code{bmixlm} Object}
\usage{
\method{coef}{bmixlm}(
  object,
  which = c("probit", "comp1", "comp2", "error"),
  type = c("mean", "samples"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{which}{The parameter set to extract (see details).}

\item{type}{Whether to return the posterior mean or samples from the posterior.}

\item{...}{Currently unused.}
}
\value{
If \code{type="mean"} the coefficients are returned as a
vector, and if \code{type="samples"} the coefficients are
returned as an array of samples from the posterior
}
\description{
Extract a set of model coefficients from a fitted \code{bmixlm} object.
}
\details{
The fitted object contains of four sets of parameters: the
coefficients for the two component models, the coefficients for
the probit model, and the error standard deviations from the two
component models. The \code{which} argument determines which of
these parameter sets is returned:
\itemize{
\item \strong{"comp1"}: coefficients for the first component model
\item \strong{"comp2"}: coefficients for the second component model
\item \strong{"probit"}: coefficients for the probit binary model
\item \strong{"error"}: error standard deviations for the two components
}
}
