% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{fitted.bmixlm}
\alias{fitted.bmixlm}
\title{Fitted Values for a \code{bmixlm} Object}
\usage{
\method{fitted}{bmixlm}(object, type = c("mean", "samples"), ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{type}{Whether to return the posterior mean or samples from
the posterior.}

\item{...}{Currently unused.}
}
\value{
Returns a list or datafame with elements
\item{\code{y1}}{predicted values for the first component model}
\item{\code{y2}}{predicted values for the second component model}
\item{\code{p}}{probabilities of membership of the second component}
If \code{type="mean"} return a datafame of posterior means is
returned, and if \code{type="samples"} return a list of arrays of
samples from the posterior.
}
\description{
Calculate the fitted values from the two component models, and the
probabilities of membership of the second component.
}
\details{
If \code{type="mean"} the function returns posterior mean
quantities as a dataframe, otherwise it returns samples from the
posterior as a list of arrays.
}

