% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{model.matrix.bmixlm}
\alias{model.matrix.bmixlm}
\title{Model matrices for a \code{bmixlm} Object}
\usage{
\method{model.matrix}{bmixlm}(object, which = c("probit", "comp1", "comp2"), data = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{which}{The parameter set to extract (see details).}

\item{data}{A data frame or \code{NULL}.}

\item{...}{Currently unused.}
}
\value{
The model matrix for the selected model component.
}
\description{
Extract a model matrix for one component of a \code{bmixlm} object.
}
\details{
A \code{bmixlm} object depends upon three model matrices: the
matrices for the two component models and the model matrix for
the probit model. The \code{which} argument determines which of
these matrices is returned:
\itemize{
\item \strong{"comp1"}: the model matrix for the first component model
\item \strong{"comp2"}: the model matrix for the second component model
\item \strong{"probit"}: the model matrix for the probit binary model
}

If a dataframe is supplied via the \code{data} argument, it is
used to construct the model matrix, otherwise the model matrix is
constructed from the data used to generate the fitted object
}
