% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{predict.bmixlm}
\alias{predict.bmixlm}
\title{Predicted Values for a \code{bmixlm} Object}
\usage{
\method{predict}{bmixlm}(object, newdata = NULL, type = c("mean", "samples"), ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{newdata}{A dataframe for which predictions are required}

\item{type}{Whether to return the posterior mean or samples
from the posterior.}

\item{...}{Currently unused.}
}
\value{
Returns a list or dataframe with elements
\itemize{
\item \strong{y1}: predicted values for the first component model
\item \strong{y2}: predicted values for the second component model
\item \strong{p}: posterior predictive probabilities of membership
of the second component
}

If \code{type="mean"} return a dataframe of posterior means is
returned, and if \code{type="samples"} return a list of arrays of
samples from the posterior.
}
\description{
Calculate the predicted values that can be computed from a
fitted model.
}
\details{
This function returns the predicted values conditional on the
original observed data and the prediction covariates only. An
expanded range of predictions that include quantities conditional
on the observed responses in the prediction data can be calculated
with \code{predictAll}.

If \code{type="mean"} the function returns posterior mean
quantities as a dataframe, otherwise it returns samples from the
posterior as a list of arrays.
}
