% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{print.bmixlm}
\alias{print.bmixlm}
\title{Printing bmixlm fits}
\usage{
\method{print}{bmixlm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bmixlm}}

\item{...}{Currently ignored}
}
\description{
Print method for objects of class \code{bmixlm}
}
