% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{residuals.bmixlm}
\alias{residuals.bmixlm}
\title{Residuals for a \code{bmixlm} Object}
\usage{
\method{residuals}{bmixlm}(object, type = c("mean", "samples"),
  standardize = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{type}{Whether to return the posterior mean or samples from
the posterior.}

\item{standardize}{Whether to standardize the residuals.}

\item{...}{Currently unused.}
}
\value{
Returns a list or dataframe with elements
\item{\code{y1}}{residuals for the first component model}
\item{\code{y2}}{residuals for the second component model}
\item{\code{p}}{posterior predictive probabilities of membership
of the second component}
\item{\code{q}}{posterior probabilities of membership of the
second component}
\item{\code{b}}{binary indicators of membership of the second component,
conditional on the observed response}
If \code{type="mean"} return a dataframe of posterior means is
returned, and if \code{type="samples"} return a list of arrays of
samples from the posterior.
}
\description{
Calculate the residuals from the two component models, and the
probabilities of membership of the second component.
}
\details{
If \code{type="mean"} the function returns posterior mean
quantities as a dataframe, otherwise it returns samples from the
posterior as a list of arrays.
}

