% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{simulate.bmixlm}
\alias{simulate.bmixlm}
\title{Simulate Responses from a \code{bmixlm} Object}
\usage{
\method{simulate}{bmixlm}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}.}

\item{nsim}{Number of response vectors to simulate.}

\item{seed}{An object specifying if and how the random number
generator should be initialized.}

\item{...}{Additional arguments to be passed to \code{update}.}
}
\value{
A dataframe of simulated responses.
}
\description{
New observations are simulated from the posterior predictive
distribution.
}
\details{
If \code{nsim} is not \code{NULL}, \code{update} is called to
generate \code{nsim} new samples from the posterior, and new
observations are generated from these. Otherwise new observations
are generated from the samples contained in the fitted object.
}
