% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{summary.bmixlm}
\alias{print.summary.bmixlm}
\alias{summary.bmixlm}
\title{Summarizing \code{bmixlm} Fits}
\usage{
\method{summary}{bmixlm}(object, ...)

\method{print}{summary.bmixlm}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{object}{An object of class \code{\link{bmixlm}}}

\item{...}{Currently ignored}

\item{x}{An object of class \code{\link{bmixlm}}}

\item{digits}{The number of significant digits to use when printing.}
}
\value{
Returns an object of class \code{summary.bmixlm}, with components
\item{\code{call}}{The original \code{bmixlm} call}
\item{\code{nsamp}}{The number of Gibbs samples drawn}
\item{\code{beta1}}{Summary table for the coefficients for the linear model for the first component}
\item{\code{beta2}}{Summary table for the coefficients for the linear model for the second component}
\item{\code{betap}}{Summary table for the coefficients for the probit model}
\item{\code{sigma}}{Summary table for the standard deviations for the two linear models}
\item{\code{vcov1}}{Variance covariance matrix of the coefficients for the linear model for the first component}
\item{\code{vcov2}}{Variance covariance matrix of the coefficients for the linear model for the second component}
\item{\code{vcovp}}{Variance covariance matrix of the coefficients for the probit model}
\item{\code{pW}}{WAIC effective number of parameters}
\item{\code{WAIC}}{WAIC}
}
\description{
Computes a list of summary statistics for the \code{bmixlm} model
fit \code{object}.
}

