% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{update.bmixlm}
\alias{update.bmixlm}
\title{Update and Refit a \code{bmixlm} Model}
\usage{
\method{update}{bmixlm}(object, ..., evaluate = TRUE)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{...}{Arguments to update}

\item{evaluate}{If true evaluate the new call else return the call.}
}
\value{
If \code{evaluate = TRUE} the fitted object, otherwise the updated call.
}
\description{
Updates and optionally refits a \code{bmixlm} fit.
}
\details{
If the \code{betap} and \code{sigma} arguments are not specified,
they are determined from the existing fit.
}
