% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bmixlm-package}
\alias{bmixlm-package}
\title{Binary Mixtures of Linear Models}
\description{
Fit a binary mixture of linear models by Gibbs sampling.  Assumes
responses are drawn two subpopulations, each described by a
different linear model, and it is unknown which responses are drawn
from which subpopulation.  The mixture model estimates the
coefficient of the two subpopulation models and the probability
that an observation is drawn from a particular subpopulation.
}
\author{
Simon Wotherspoon
}

