% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixlm.R
\name{summary.bmixlm}
\alias{summary.bmixlm}
\alias{print.summary.bmixlm}
\title{Summarizing \code{bmixlm} Fits}
\usage{
\method{summary}{bmixlm}(object, ...)

\method{print}{summary.bmixlm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object of class \code{bmixlm}}

\item{...}{Currently ignored}

\item{x}{An object of class \code{bmixlm}}

\item{digits}{The number of significant digits to use when printing.}
}
\value{
Returns an object of class \code{summary.bmixlm}, with components
\itemize{
\item \strong{call}: The original \code{bmixlm} call
\item \strong{nsamp}: The number of Gibbs samples drawn
\item \strong{beta1}: Summary table for the coefficients for the linear model for the first component
\item \strong{beta2}: Summary table for the coefficients for the linear model for the second component
\item \strong{betap}: Summary table for the coefficients for the probit model
\item \strong{sigma}: Summary table for the standard deviations for the two linear models
\item \strong{vcov1}: Variance covariance matrix of the coefficients for the linear model for the first component
\item \strong{vcov2}: Variance covariance matrix of the coefficients for the linear model for the second component
\item \strong{vcovp}: Variance covariance matrix of the coefficients for the probit model
\item \strong{pW}: WAIC effective number of parameters
\item \strong{WAIC}: WAIC
}
}
\description{
Computes a list of summary statistics for the \code{bmixlm} model
fit \code{object}.
}
